Move params method to common super class
