+++ client/src/components/city.js
@@ -1,4 +1,6 @@
-import React, {PureComponent} from 'react';
+import React from 'react';
+import CityBase from './city-base';
+
 import {browserHistory} from 'react-router';
 import PropTypes from 'prop-types';
 
@@ -16,7 +18,7 @@ import CityStore from '../stores/city-store';
 import CityViewStore from '../stores/city-view-store';
 import EditorStore from '../stores/editor-store';
 
-class City extends PureComponent {
+class City extends CityBase {
   constructor(props, context) {
     super(props, context);
 
@@ -65,23 +67,6 @@ class City extends PureComponent {
       MainStore.unsetLoading();
       browserHistory.push(`/error?path=${this.urlName}`);
     }
-  }
-
-  params() {
-    return this.props.location.query;
-  }
-
-  updateParams(newParams) {
-    const params = Object.assign({}, this.params(), newParams);
-
-    // We delete null params
-    Object.keys(params).forEach((key) => (params[key] == null) && delete params[key]);
-
-    // If new params are equal to the current ones, we don't push the state to the
-    // browser history
-    if (JSON.stringify(params) === JSON.stringify(this.params())) return;
-
-    browserHistory.push({...this.props.location, query: params});
   }
 
   onChange() {
